feat: refactor signature feature
create new class to handle signature results after signing a tx
